Migrate from Integer to BigInteger for UserIDs
Query to migrate the database itself:
ALTER TABLE afk_users ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE userinfo ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE userbio ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE antiflood ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE gbans ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE warns ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE connection ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE connection_history ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE human_checks ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE approval ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE users ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE chat_members ALTER COLUMN priv_chat_id SET DATA TYPE BigInt;
ALTER TABLE chat_members ALTER COLUMN "user" SET DATA TYPE BigInt;
ALTER TABLE royals ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE user_report_settings ALTER COLUMN user_id SET DATA TYPE BigInt;

+++ tg_bot/modules/sql/connection_sql.py
@@ -3,6 +3,7 @@ import time
 from typing import Union
 
 from sqlalchemy import Column, String, Boolean, UnicodeText, Integer
+from sqlalchemy.sql.sqltypes import BigInteger
 
 from tg_bot.modules.sql import SESSION, BASE
 
@@ -24,7 +25,7 @@ class ChatAccessConnectionSettings(BASE):
 
 class Connection(BASE):
     __tablename__ = "connection"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     chat_id = Column(String(14))
 
     def __init__(self, user_id, chat_id):
@@ -34,7 +35,7 @@ class Connection(BASE):
 
 class ConnectionHistory(BASE):
     __tablename__ = "connection_history"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     chat_id = Column(String(14), primary_key=True)
     chat_name = Column(UnicodeText)
     conn_time = Column(Integer)
